+++ src/main/java/com/noradltd/wumpus/OccupantManager.java
@@ -1,11 +1,15 @@
 package com.noradltd.wumpus;
 
+import java.io.Serial;
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-class OccupantManager {
+class OccupantManager implements Serializable {
+    @Serial
+    private static final long serialVersionUID = 1L;
 
     private final List<Room.Occupant> occupants = new ArrayList<>();